e2e/gateway: extract InitTF and rename Describe files for cross-package import
Blank-importing from a parent test binary requires Describes in regular
.go files (Go excludes _test.go from cross-package imports). Each suite
exposes state.go with InitTF() and moves specs to non-test files.
