+++ test/e2e/gateway/chained_gateway_tests/chained_gateway_suite_test.go
@@ -3,13 +3,9 @@ package chained_gateway_tests
 import (
 	"testing"
 
-	"sigs.k8s.io/aws-load-balancer-controller/v3/test/framework"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-var tf *framework.Framework
 
 func TestChainedGateway(t *testing.T) {
 	RegisterFailHandler(Fail)
@@ -17,7 +13,5 @@ func TestChainedGateway(t *testing.T) {
 }
 
 var _ = BeforeSuite(func() {
-	var err error
-	tf, err = framework.InitFramework()
-	Expect(err).NotTo(HaveOccurred())
+	Expect(InitTF()).To(Succeed())
 })
