+++ test/e2e/gateway/listenerset_tests/gateway_listenerset_suite_test.go
@@ -3,13 +3,9 @@ package listenerset_tests
 import (
 	"testing"
 
-	"sigs.k8s.io/aws-load-balancer-controller/v3/test/framework"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-var tf *framework.Framework
 
 func TestGatewayListenerSet(t *testing.T) {
 	RegisterFailHandler(Fail)
@@ -17,7 +13,5 @@ func TestGatewayListenerSet(t *testing.T) {
 }
 
 var _ = BeforeSuite(func() {
-	var err error
-	tf, err = framework.InitFramework()
-	Expect(err).NotTo(HaveOccurred())
+	Expect(InitTF()).To(Succeed())
 })
